refactor: 공통 컴포넌트 components/common/ 으로 이동
